Moved static methods from SymbolValue to SymbolValueOps
