+++ src/main/java/io/github/rodyamirov/tree/BinOpNode.java
@@ -4,6 +4,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import io.github.rodyamirov.lex.Token;
 import io.github.rodyamirov.symbols.SymbolValue;
+import io.github.rodyamirov.symbols.SymbolValueOps;
 
 import java.util.Objects;
 import java.util.function.BiFunction;
@@ -19,22 +20,22 @@ public final class BinOpNode extends ExpressionNode {
     // static it up
     private static final ImmutableMap<Token.Type, BiFunction<SymbolValue, SymbolValue, SymbolValue>> evaluations =
             ImmutableMap.<Token.Type, BiFunction<SymbolValue, SymbolValue, SymbolValue>>builder() // java 7 fail :(
-                    .put(Token.Type.PLUS, SymbolValue::add)
-                    .put(Token.Type.MINUS, SymbolValue::subtract)
-                    .put(Token.Type.TIMES, SymbolValue::multiply)
-                    .put(Token.Type.INT_DIVIDE, SymbolValue::intDivide)
-                    .put(Token.Type.REAL_DIVIDE, SymbolValue::realDivide)
-                    .put(Token.Type.MOD, SymbolValue::intMod)
+                    .put(Token.Type.PLUS, SymbolValueOps::add)
+                    .put(Token.Type.MINUS, SymbolValueOps::subtract)
+                    .put(Token.Type.TIMES, SymbolValueOps::multiply)
+                    .put(Token.Type.INT_DIVIDE, SymbolValueOps::intDivide)
+                    .put(Token.Type.REAL_DIVIDE, SymbolValueOps::realDivide)
+                    .put(Token.Type.MOD, SymbolValueOps::intMod)
 
-                    .put(Token.Type.AND, SymbolValue::and)
-                    .put(Token.Type.OR, SymbolValue::or)
+                    .put(Token.Type.AND, SymbolValueOps::and)
+                    .put(Token.Type.OR, SymbolValueOps::or)
 
-                    .put(Token.Type.LESS_THAN, SymbolValue::lessThan)
-                    .put(Token.Type.LESS_THAN_OR_EQUALS, SymbolValue::lessThanOrEquals)
-                    .put(Token.Type.GREATER_THAN, SymbolValue::greaterThan)
-                    .put(Token.Type.GREATER_THAN_OR_EQUALS, SymbolValue::greaterThanOrEquals)
-                    .put(Token.Type.EQUALS, SymbolValue::equalsValue)
-                    .put(Token.Type.NOT_EQUALS, SymbolValue::notEqualsValue)
+                    .put(Token.Type.LESS_THAN, SymbolValueOps::lessThan)
+                    .put(Token.Type.LESS_THAN_OR_EQUALS, SymbolValueOps::lessThanOrEquals)
+                    .put(Token.Type.GREATER_THAN, SymbolValueOps::greaterThan)
+                    .put(Token.Type.GREATER_THAN_OR_EQUALS, SymbolValueOps::greaterThanOrEquals)
+                    .put(Token.Type.EQUALS, SymbolValueOps::equalsValue)
+                    .put(Token.Type.NOT_EQUALS, SymbolValueOps::notEqualsValue)
                     .build();
     private static final ImmutableSet<Token.Type> allowedOpTypes = evaluations.keySet();
