+++ src/main/java/io/github/rodyamirov/utils/Procedure.java
@@ -1,0 +1,8 @@
+package io.github.rodyamirov.utils;
+
+/**
+ * Created by richard.rast on 12/30/16.
+ */
+public interface Procedure {
+    void invoke();
+}
